end api task list & tasks & task_list_user

+++ backend/app/Services/Api/TaskListService.php
@@ -4,98 +4,127 @@ namespace App\Services\Api;
 
 use App\Http\Requests\Api\TaskListPostRequest;
 use App\Models\Api\TaskList;
+use App\Models\Api\TaskListUser;
+use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Validator;
 
-class TaskListService
+class TaskListService extends BaseApiService
 {
     public function getTasksList($request): \Illuminate\Http\JsonResponse
     {
         $page       = (int)$request->query('page', 1);
         $limit      = (int)$request->query('limit', 10);
 
-        $query      = TaskList::query();
-
-        $total      = $query->count();
-        $totalPages = ceil($total / $limit);
-
-        $data       = collect([]);
-
-        $task_list  = $query
-            ->skip(($page - 1) * $limit)
+        $dataTaskListUser = TaskListUser::query()
+            ->select('fk_task_list')
             ->where([
-                'fk_user'   => auth()->id()
+                'fk_user' => auth()->id(),
             ])
+            ->skip(($page - 1) * $limit)
+            ->take($limit)
+            ->pluck('fk_task_list');
+
+        if($dataTaskListUser->count() === 0){
+            return response()->json([
+                'success'   => true,
+                'data'      => [
+                    'data'              => $dataTaskListUser,
+                    'current_page'      => $page,
+                    'limit'             => $limit,
+                    'total_records'     => 0,
+                    'total_pages'       => 0,
+                ],
+            ], 200);
+        }
+
+        $taskLists = TaskList::query()
+            ->whereIn('id', $dataTaskListUser)
+            ->skip(($page - 1) * $limit)
             ->take($limit)
             ->get();
 
-        foreach ($task_list as $task){
-            $newData = [
+        $data = $taskLists->map(function ($task) {
+
+            $userWithPermission = $task->usersWithPermission->first();
+
+            return [
                 'id'            => $task->id,
                 'name'          => $task->name,
                 'description'   => $task->description,
-                'is_completed'  => $task->iscompleted,
+                'is_completed'  => $task->is_completed,
                 'created_at'    => $task->created_at->format('d/m/Y H:i'),
                 'updated_at'    => $task->updated_at->format('d/m/Y H:i'),
-                'user'          => $task->fkUser->name
+                'user'          => $userWithPermission->name ?? null,
+                'permission'    => $userWithPermission->pivot->permission ?? null,
             ];
-
-            $data[] = $newData;
-        }
+        });
 
         return response()->json([
-            'success'   => true,
-            'data'      => [
-                'data'              => $data,
-                'current_page'      => $page,
-                'limit'             => $limit,
-                'total_records'     => $total,
-                'total_pages'       => $totalPages,
+            'success' => true,
+            'data'    => [
+                'data'          => $data,
+                'current_page'  => $page,
+                'limit'         => $limit,
+                'total_records' => $taskLists->count(),
+                'total_pages'   => ceil($taskLists->count() / $limit),
             ],
         ], 200);
+
     }
 
     public function createTaskList($taskList): \Illuminate\Http\JsonResponse
     {
-        $request = new TaskListPostRequest();
+        try {
+            DB::beginTransaction();
 
-        $rules = $request->rules();
+            $request = new TaskListPostRequest();
 
-        $validator = Validator::make($taskList->all(), $rules);
+            $rules = $request->rules();
 
-        if($validator->fails()){
+            $validator = Validator::make($taskList->all(), $rules);
+
+            if ($validator->fails()) {
+                return response()->json([
+                    'success' => false,
+                    'message' => $validator->errors()->getMessages()
+                ], 400);
+            }
+
+            $taskListModel = new TaskList();
+
+            $taskListModel->fill([
+                'name'          => $taskList->name,
+                'description'   => $taskList->description,
+                'is_completed'  => $taskList->is_completed,
+                'fk_user'       => auth()->id()
+            ]);
+
+            $taskListModel->save();
+
+            DB::commit();
+
             return response()->json([
-                'success'   => false,
-                'message'   => $validator->errors()->getMessages()
-            ], 400);
+                'success' => true,
+                'data' => [
+                    'id' => $taskListModel->id
+                ]
+            ], 201);
+
+        }catch (\Exception $exception){
+            DB::rollback();
+
+            return response()->json([
+                'success' => false,
+                'message' => $exception->getMessage()
+            ],200);
         }
-
-        $taskListModel = new TaskList();
-
-        $user = auth()->user();
-
-        $taskListModel->fill([
-            'name'          => $taskList->name,
-            'description'   => $taskList->description,
-            'fk_user'       => $user->id,
-            'is_completed'  => $taskList->is_completed
-        ]);
-
-        $taskListModel->save();
-
-        return response()->json([
-            'success'   => true,
-            'data'      => [
-                'id'    =>  $taskListModel->id
-            ]
-        ], 201);
     }
 
     public function showTaskList($id): \Illuminate\Http\JsonResponse
     {
         $task_list = TaskList::query()
             ->where([
-                'id'        =>  $id,
-                'fk_user'   => auth()->id()
+                'id'        =>  $id
             ])->first();
 
         if(!$task_list){
@@ -105,6 +134,12 @@ class TaskListService
             ], 404);
         }
 
+        $taskListToUser = $this->getUserPermission($task_list->id, ['full', 'edit', 'view']);
+
+        if($taskListToUser){
+            return $taskListToUser;
+        }
+
         return response()->json([
             'success'   => true,
             'message'   => 'show task',
@@ -112,9 +147,9 @@ class TaskListService
                 'id'            => $task_list->id,
                 'name'          => $task_list->name,
                 'description'   => $task_list->description,
-                'user'          => $task_list->fkUser->name,
                 'created_at'    => $task_list->created_at->format('d/m/Y H:i'),
-                'is_completed'  => $task_list->is_completed
+                'is_completed'  => $task_list->is_completed,
+                'user'          => $taskListToUser->user->name
             ]
         ], 200);
     }
@@ -122,8 +157,7 @@ class TaskListService
     public function updatedTaskList($taskList, $id): \Illuminate\Http\JsonResponse
     {
         $taskListModel = TaskList::query()->where([
-            'id'        => $id,
-            'fk_user'   => auth()->id()
+            'id'        => $id
         ])->first();
 
         if(!$taskListModel){
@@ -131,6 +165,12 @@ class TaskListService
                 'success'   => false,
                 'message'   => 'Not found'
             ], 404);
+        }
+
+        $taskListToUser = $this->getUserPermission($taskListModel->id, ['full', 'edit']);
+
+        if($taskListToUser){
+            return $taskListToUser;
         }
 
         $request = new TaskListPostRequest();
@@ -159,20 +199,26 @@ class TaskListService
             'message'   => 'update task',
             'data'      => $id
         ], 200);
+
     }
 
     public function destroyTaskList($id): \Illuminate\Http\JsonResponse
     {
         $taskListModel = TaskList::query()->where([
-            'id'        => $id,
-            'fk_user'   => auth()->id()
+            'id'        => $id
         ])->first();
 
         if(!$taskListModel){
             return response()->json([
                 'success'   => false,
-                'message'   => 'Not found'
+                'message'   => 'Not found!'
             ], 404);
+        }
+
+        $taskListToUser = $this->getUserPermission($taskListModel->id, ['full']);
+
+        if($taskListToUser){
+            return $taskListToUser;
         }
 
         $taskListModel->delete();
